prepare_data/requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
certifi==2025.6.15
charset-normalizer==3.4.2
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
easydict==1.13
flatbuffers==25.2.10
fonttools==4.58.4
humanfriendly==10.0
idna==3.10
imageio==2.37.0
insightface==0.7.3
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
onnx==1.18.0
onnxruntime==1.22.0
onnxruntime-gpu==1.22.0
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.2.1
prettytable==3.16.0
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.16.0
setuptools==78.1.1
simsimd==6.4.9
six==1.17.0
stringzilla==3.12.5
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.6.11
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
wheel==0.45.1
